[MTV-2926] address feedback
